+++ src/main/java/edu/estu/Knapsack.java
@@ -2,9 +2,9 @@ package edu.estu;
 
 import java.util.List;
 
-public record Knapsack(int capacity, List<Item> items) {
+public record Knapsack(long capacity, List<Item> items) {
     public boolean isTrivial() {
-        int totalWeight = items.stream().mapToInt(Item::weight).sum();
+        long totalWeight = items.stream().mapToLong(Item::weight).sum();
         return totalWeight <= capacity;
     }
 }
